Crate gamepower_wallet

Re-exports

pub use $crate::traits::GetPalletVersion as _;

Structs

AllClaims
Get a vector of all claims. Used as a quick lookup.

AllListings
Get a vector of all listings. Used as a quick lookup.

Claim
Claim data

Emotes
Emotes used by the wallet

Listing
Listing data

ListingCount
A fast and simple count of all current listings

Listings
Get a listing by the listing_id

ListingsByOwner
Get all listings ids by an account

Module

NextClaimId
Get the next claim id

NextListingId
Get the next listing id

OpenClaims
Get one or more claims by AccountId or a single claim including the claim_id

Order
Order data

OrderCount
A count of all orders made through the wallet

OrderHistory
A history of orders for an asset

Enums

Call
Dispatchable calls.

Error

RawEvent
Events for this module.

Traits

Config
The module configuration trait.

Type Definitions

ClaimOf
Claim Data

ClassIdOf
Class Id

Event
RawEvent specialized for the configuration Config

ListingOf
Listing Data

OrderOf
Order Data

TokenIdOf
Token Id